requirements.txt
streamlit==1.32.0
faiss-cpu==1.8.0
sentence-transformers==2.7.0
ollama<0.1.14
langchain==0.1.16
rank-bm25==0.2.2
python-dotenv==1.0.1
google-search-results==2.4.2
rapidfuzz
bs4